refactor: LotteryResult 로 통합
로또 결과 객체에서 모든 결과 관련 로직을 수행하도록 수정
